add validation to api

+++ app/Http/Controllers/StudentController.php
@@ -4,6 +4,8 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Student;
+use DateTime;
+use Illuminate\Support\Facades\Date;
 
 class StudentController extends Controller
 {
@@ -14,20 +16,31 @@ class StudentController extends Controller
 
     public function create(Request $request)
     {
+        date_default_timezone_set('Asia/Jakarta');
+        $timestamp = date('Y-m-d H:i:s');
+
         $student = new Student();
 
         $check = Student::where('name', $request->name)->first();
+
+        if ($request->name == null || $request->nim == null || $request->email == null || $request->jurusan == null) {
+            $data = [
+                'message' => 'parameter cannot be empty'
+            ];
+
+            return response()->json($data);
+        }
         
         if ($check != null) {
             $data = [
-                'message' => 'name already exists',
-                'data' => null
+                'message' => 'name already exists'
             ];
         } else {
             $student->name = $request->name;
             $student->nim = $request->nim;
             $student->email = $request->email;
             $student->jurusan = $request->jurusan;
+            $student->created_at = $timestamp;
             $student->save();
             
             $data = [
@@ -41,16 +54,46 @@ class StudentController extends Controller
 
     public function find(Request $request, $id)
     {
-        return response()->json(Student::where('id', $id)->first());
+        if (!is_numeric($id)) {
+            $data = [
+                'message' => 'data must be number'
+            ];
+
+            return response()->json($data, 503);
+        }
+
+        $find = Student::where('id', $id)->first();
+        if ($find == null) {
+            $data = [
+                'messege' => 'data not found'
+            ];
+
+            return response()->json($data);
+        } else {
+            return response()->json($find);
+        }
     }
 
     public function update(Request $request)
     {
+        date_default_timezone_set('Asia/Jakarta');
+        $timestamp = date('Y-m-d H:i:s');
+
         $doChange = Student::where('id', $request->id)->first();
-        $doChange->name = $request->name;
-        $doChange->nim = $request->nim;
-        $doChange->email = $request->email;
-        $doChange->jurusan = $request->jurusan;
+
+        if ($doChange == null) {
+            $data = [
+                'message' => 'the data you want to update was not found'
+            ];
+
+            return response()->json($data, 404);
+        }
+
+        $doChange->name = $request->name != null ? $request->name : $doChange->name;
+        $doChange->nim = $request->nim != null ? $request->nim : $doChange->nim;
+        $doChange->email = $request->email != null ? $request->email : $doChange->email;
+        $doChange->jurusan = $request->jurusan != null ? $request->jurusan : $doChange->jurusan;
+        $doChange->updated_at = $timestamp;
         $doChange->save();
 
         $message = [
@@ -65,6 +108,14 @@ class StudentController extends Controller
     {
         $delete = Student::where('id', $request->id)->first();
         
+        if ($delete == null) {
+            $data = [
+                'message' => 'the data you want to delete was not found'
+            ];
+
+            return response()->json($data, 404);
+        }
+
         $delete->delete();
 
         $message = [
@@ -72,6 +123,6 @@ class StudentController extends Controller
             'data' => $delete
         ];
 
-        return response()->json($message, 201);
+        return response()->json($message, 200);
     }
 }
